Added way to init Vulkan by passing in vkGetInstanceProcAddr func pointer.
For some reason when we do the archive build, which is required for releasing
an app on the Apple App Store, the dlsym(RTLD_DEFAULT, "vkGetInstanceProcAddr")
call fails, which causes SDL_Vulkan_LoadLibrary to fail. To work around that
issue I added SDL_Vulkan_Init which takes a pointer to the vkGetInstanceProcAddr
function, which on iOS is not dynamically loaded anyway, and as such does not
need to be searched for via dlsym.

This seems like a viable work around until we figure out and fix the issue
with the XCode archive build.