Fixed a deadlock when closing the audio device.
The audio player thread was getting stuck waiting for a semaphore to be incremented
when the SDL_CloseAudioDevice function was called. The easiest fix was to switch to
a timed wait on the semaphore that would check if shutdown had been called if the
wait on the semaphore timed out.