Add define SDL_IOS_EMBEDDED to exclude main() for tvOS (#6)
Had to use of the SDL_IOS_EMBEDDED macro, which had previously been added to prevent inclusion of the main function for when we use SDL on iOS devices, because it causes a duplicate symbol error due to the fact that we aren't using SDL like a normal SDL app would use SDL on an iOS or tvOS device. That fact combined with the fact that Flutter forces us to use the -ObjC linker flag which forces all linked libraries to load all their symbols even if they aren't explicitly needed requires us to have this extra macro to remove the main() that SDL adds.